Refactor code to make it more coherent
- Move utils into lib (as it is done in sprotty)
- Rename listener that moves the feedback end
- Enable and disable listener only when tool is active
