+++ client/packages/glsp-sprotty/src/lib/utils/operation.ts
