+++ client/packages/glsp-sprotty/src/lib/utils/smodel-util.ts
