+++ client/packages/glsp-sprotty/src/lib/utils/viewpoint-util.ts
@@ -1,0 +1,49 @@
+/*******************************************************************************
+ * Copyright (c) 2018 EclipseSource Services GmbH and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Contributors:
+ *  Philip Langer - initial API and implementation
+ ******************************************************************************/
+
+import { findParentByFeature, isViewport, Point, SModelElement, Viewport } from "sprotty/lib";
+
+/**
+ * Return the position corresponding to this mouse event (Browser coordinates)
+ * in the diagram coordinates system (i.e. relative to the Diagram's 0;0 point)
+ *
+ * This functions takes into account the following transformations:
+ * - Location of the Diagram Canvas inside of the browser's page
+ * - Current viewport Scroll and Zoom
+ *
+ * @param target
+ *  An element from the diagram
+ * @param mouseEvent
+ *  A mouseEvent
+ */
+export function getAbsolutePosition(target: SModelElement, mouseEvent: MouseEvent): Point {
+    let xPos = mouseEvent.pageX, yPos = mouseEvent.pageY;
+    const canvasBounds = target.root.canvasBounds;
+    xPos -= canvasBounds.x;
+    yPos -= canvasBounds.y;
+
+    const viewport: Viewport | undefined = findParentByFeature(target, isViewport);
+    const zoom = viewport ? viewport.zoom : 1;
+    if (viewport) {
+        const scroll: Point = { x: viewport.scroll.x, y: viewport.scroll.y }
+        xPos += scroll.x * zoom;
+        yPos += scroll.y * zoom;
+
+        xPos /= zoom;
+        yPos /= zoom;
+    }
+    xPos
+
+    return {
+        x: xPos,
+        y: yPos
+    }
+}
